Resolved URL issues, Removed Excel Interop
I had issues reading files from the URL /Content/Uploads when deploying
to Web Server because of my use of a virtual Directory that points to an
alternative location on the file system.  I removed the filesystem path
to the file so that I could protect the directory structure on the Web
server.  I removed the Excel Interop assembly references because I
needed MS Office installed on the server in order to use them.  So I
will work to Utilize Open XML instead.

+++ MvcPlanningApplication/Models/HaworthSupplierDemand.cs
@@ -49,9 +49,9 @@ namespace MvcPlanningApplication.Models
         public string DelAppt { get; set; }
         public string DelSite { get; set; }
         public int Route { get; set; }
-        public int NetPrice { get; set; }
-        public int NetValue { get; set; }
-        public int POTotal { get; set; }
+        public double NetPrice { get; set; }
+        public double NetValue { get; set; }
+        public double POTotal { get; set; }
         public string SrcLoc { get; set; }
     }
 }
